creating service folder, using axios instance as api
