+++ web/src/components/AlertListItem/index.js
@@ -1,7 +1,7 @@
 import React, { Component } from 'react';
 import { bindActionCreators } from 'redux';
 import { connect } from 'react-redux';
-import { selectAlert, deleteAlert } from '../../actions/Alert';
+import { selectAlert, deleteAlert } from '../../service/Alert';
 import Swal from 'sweetalert2'
 
 class AlertListItem extends Component {
