--- web/src/configs/index.js
@@ -1,1 +1,0 @@
-export const API_URL = 'http://localhost:3001';
