+++ web/src/service/api.js
@@ -1,0 +1,7 @@
+import axios from 'axios'
+
+const api = axios.create({
+    baseURL: 'http://localhost:3001/'
+})
+
+export default api
